fix: add missing user.api.ts and resolve import path errors
- Create src/app/api/user.api.ts with fetchUserProfile and fetchUserReputation
- Refactor user.queries.ts to import from user.api.ts (avoid duplication)
- Fix import paths in useUser.ts, useClaims.ts, and useLeaderboard.ts
- Resolve build failure from missing module (issue #76)
