+++ src/app/queries/user.queries.ts
@@ -2,38 +2,10 @@
 
 import { useQuery } from '@tanstack/react-query';
 import { queryKeys } from './queryKeys';
+import { fetchUserProfile, fetchUserReputation, UserProfile, UserReputation } from '../api/user.api';
 
-export interface UserProfile {
-  id: string;
-  address: string;
-  username: string;
-  reputation: number;
-  verificationCount: number;
-  accuracy: number;
-  totalStaked: number;
-  totalEarned: number;
-  joinedAt: string;
-}
-
-export interface UserReputation {
-  score: number;
-  rank: number;
-  totalVerifications: number;
-  successfulVerifications: number;
-  accuracy: number;
-}
-
-async function fetchUserProfile(userId: string): Promise<UserProfile> {
-  const res = await fetch(`/api/users/${userId}`);
-  if (!res.ok) throw new Error('Failed to fetch user profile');
-  return res.json();
-}
-
-async function fetchUserReputation(userId: string): Promise<UserReputation> {
-  const res = await fetch(`/api/users/${userId}/reputation`);
-  if (!res.ok) throw new Error('Failed to fetch user reputation');
-  return res.json();
-}
+// Re-export types for backward compatibility
+export type { UserProfile, UserReputation };
 
 export function useUserProfile(userId: string) {
   return useQuery(
